Update and rename corrected.py to corrected_code.py
